+++ Sources/Pippin/Extensions/Foundation/CGFloat/CGFloat+ThemeConstants.swift
@@ -1,0 +1,12 @@
+//
+//  CGFloat+ThemeConstants.swift
+//  Pippin
+//
+//  Created by Andrew McKnight on 1/24/19.
+//
+
+import Foundation
+
+public extension CGFloat {
+    static let disabledAlpha: CGFloat = 0.6
+}